Updated makexpi.sh

+++ makexpi.sh
@@ -1,6 +1,11 @@
 #!/bin/sh
 #
-# Usage: makexpi.sh -n <addonname> -v <flag> -s <suffix>
+# Usage: makexpi.sh -n <addonname> -v -s <suffix> -o
+#
+#          -v : generate XPI file with version number
+#          -f : generate only XPI (high complression)
+#          -o : generate only XPI (no complression, omnixpi)
+#
 #        ex.
 #         $ ./makexpi.sh -n myaddon -v 1
 #         $ ./makexpi.sh -n myaddon
@@ -48,12 +53,18 @@
 #          + [skin]
 
 
-while getopts n:v:s: OPT
+use_version=0
+nojar=0
+xpi_compression_level=9
+
+while getopts n:os:v OPT
 do
   case $OPT in
-    "n" ) appname="$OPTARG";;
-    "v" ) use_version="$OPTARG"; use_version=`echo "$use_version" | sed -r -e 's#version=(1|yes|true)#1#ig'`;;
+    "n" ) appname="$OPTARG" ;;
+    "v" ) use_version=1 ;;
     "s" ) suffix="$OPTARG" ;;
+    "o" ) nojar=1; xpi_compression_level=0 ;;
+    "f" ) nojar=1 ;;
   esac
 done
 
@@ -98,7 +109,7 @@ then
 fi
 
 
-xpi_contents="chrome components/*.js components/*.xpt modules isp defaults license platform *.js *.rdf *.manifest *.inf *.cfg *.light"
+xpi_contents="chrome components/*.js components/*.xpt components/*/*.xpt modules isp defaults license platform *.js *.rdf chrome.manifest *.inf *.cfg *.light icon*.png"
 
 
 
@@ -135,33 +146,41 @@ then
 
 	rm components/*.idl
 
-	for dirname in *
-	do
-		if [ -d $dirname/content -o -d $dirname/skin -o -d $dirname/locale ]
-		then
-			cd $dirname
-			mkdir -p chrome
-			zip -r -0 chrome/$appname.jar content locale skin -x \*/.svn/\*
-			rm -r -f content
-			rm -r -f locale
-			rm -r -f skin
-			cd ..
-		fi
-	done
+    if [ "$nojar" = "0" ]
+    then
+		for dirname in *
+		do
+			if [ -d $dirname/content -o -d $dirname/skin -o -d $dirname/locale ]
+			then
+				cd $dirname
+				mkdir -p chrome
+				zip -r -0 chrome/$appname.jar content locale skin -x \*/.svn/\*
+				rm -r -f content
+				rm -r -f locale
+				rm -r -f skin
+				cd ..
+			fi
+		done
+	fi
 	cd ../..
 fi
 
 
 cd xpi_temp
-chmod -R 644 *.jar *.js *.light *.inf *.rdf *.cfg *.manifest
+chmod -R 644 *.*
 
 
-# create jar
-mkdir -p chrome
-zip -r -0 ./chrome/$appname.jar content locale skin -x \*/.svn/\*
-if [ ! -f ./chrome/$appname.jar ]
+if [ "$nojar" = "0" ]
 then
-	rm -r -f chrome
+	# create jar
+	mkdir -p chrome
+	zip -r -0 ./chrome/$appname.jar content locale skin -x \*/.svn/\*
+	if [ ! -f ./chrome/$appname.jar ]
+	then
+		rm -r -f chrome
+	fi
+else
+	xpi_contents="content locale skin $xpi_contents"
 fi
 
 
@@ -174,12 +193,12 @@ fi
 
 
 #create xpi (Japanese)
-zip -r -9 ../$appname${version_part}${suffix}.xpi $xpi_contents -x \*/.svn/\* || exit 1
+zip -r -$xpi_compression_level ../$appname${version_part}${suffix}.xpi $xpi_contents -x \*/.svn/\* || exit 1
 
 #create xpi without update info (Japanese)
 rm -f install.rdf
 sed -e "s#^.*<em:*\(updateURL\|updateKey\)>.*</em:*\(updateURL\|updateKey\)>##g" -e "s#^.*em:*\(updateURL\|updateKey\)=\(\".*\"\|'.*'\)##g" ../install.rdf > install.rdf
-zip -r -9 ../${appname}${version_part}${suffix}_noupdate.xpi $xpi_contents -x \*/.svn/\* || exit 1
+zip -r -$xpi_compression_level ../${appname}${version_part}${suffix}_noupdate.xpi $xpi_contents -x \*/.svn/\* || exit 1
 
 
 
@@ -200,11 +219,11 @@ then
 	cp ../en.inf ./locale.inf
 	cp ../options.$appname.en.inf ./options.inf
 	chmod 644 *.inf
-	zip -r -9 ../${appname}${suffix}${version_part}_en.xpi $xpi_contents -x \*/.svn/\* || exit 1
+	zip -r -$xpi_compression_level ../${appname}${suffix}${version_part}_en.xpi $xpi_contents -x \*/.svn/\* || exit 1
 
 	rm -f install.rdf
 	sed -e "s#^.*<em:*\(updateURL\|updateKey\)>.*</em:*\(updateURL\|updateKey\)>##g" -e "s#^.*em:*\(updateURL\|updateKey\)=\(\".*\"\|'.*'\)##g" ../install.rdf > install.rdf
-	zip -r -9 ../${appname}${suffix}${version_part}_noupdate_en.xpi $xpi_contents -x \*/.svn/\* || exit 1
+	zip -r -$xpi_compression_level ../${appname}${suffix}${version_part}_noupdate_en.xpi $xpi_contents -x \*/.svn/\* || exit 1
 fi
